cloud_functions/post_processing/requirements.txt
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0